+++ Common/src/main/java/com/tristankechlo/additionalredstone/client/screen/SequencerScreen.java
@@ -1,63 +1,54 @@
 package com.tristankechlo.additionalredstone.client.screen;
 
 import com.tristankechlo.additionalredstone.Constants;
+import com.tristankechlo.additionalredstone.client.util.CustomScreen;
+import com.tristankechlo.additionalredstone.init.ModBlocks;
 import com.tristankechlo.additionalredstone.network.IPacketHandler;
+import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.components.EditBox;
-import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 
-public class SequencerScreen extends Screen {
+public class SequencerScreen extends CustomScreen {
 
-    private static final Component TITLE = Component.translatable("screen.additionalredstone.sequencer");
-    private static final ResourceLocation ERROR = new ResourceLocation(Constants.MOD_ID, "textures/gui/icons.png");
+    private static final Component TITLE = ModBlocks.SEQUENCER_BLOCK.get().getName().withStyle(ChatFormatting.BOLD);
+    private static final Component INTERVAL = Component.translatable("screen.additionalredstone.sequencer.interval");
+    private static final ResourceLocation TEXTURE = new ResourceLocation(Constants.MOD_ID, "textures/gui/sequencer_screen.png");
+    private final BlockPos pos;
     private EditBox intervalWidget;
-    private Button saveButton;
-    private Button cancelButton;
-    private BlockPos pos;
-    private int interval;
+    private final int initialInterval;
     private boolean intervalError = false;
 
     public SequencerScreen(int interval, BlockPos pos) {
-        super(TITLE);
-        this.interval = interval;
+        super(TITLE, 256, 86);
+        this.initialInterval = interval;
         this.pos = pos;
-    }
-
-    @Override
-    public void tick() {
-        this.intervalWidget.tick();
     }
 
     @Override
     protected void init() {
         super.init();
-        this.intervalWidget = new EditBox(this.font, this.width / 2 + 32, 60, 98, 20, Component.translatable("screen.additionalredstone.sequencer.interval"));
-        this.addWidget(this.intervalWidget);
+        this.intervalWidget = new EditBox(this.font, this.leftPos + 176, this.topPos + 24, 70, 20, INTERVAL);
         this.intervalWidget.setMaxLength(10);
-        this.setInitialFocus(this.intervalWidget);
-        this.intervalWidget.setFocused(true);
-        this.intervalWidget.setValue(String.valueOf(this.interval));
+        this.intervalWidget.setValue(String.valueOf(this.initialInterval));
+        this.addRenderableWidget(this.intervalWidget);
 
-        this.saveButton = new Button.Builder(Component.translatable("screen.additionalredstone.save"), (b) -> this.save())
-                .pos(this.width / 2 - 110, 150).size(100, 20).build();
-        this.cancelButton = new Button.Builder(Component.translatable("screen.additionalredstone.cancel"), (b) -> this.cancel())
-                .pos(this.width / 2 + 10, 150).size(100, 20).build();
+        Button saveButton = new Button.Builder(CustomScreen.TEXT_SAVE, this::save)
+                .pos(this.leftPos + 9, this.topPos + 57).size(116, 20)
+                .tooltip(CustomScreen.TOOLTIP_SAVE.get()).build();
+        Button cancelButton = new Button.Builder(CustomScreen.TEXT_CANCEL, (b) -> this.onClose())
+                .pos(this.leftPos + 131, this.topPos + 57).size(116, 20)
+                .tooltip(CustomScreen.TOOLTIP_CANCEL.get()).build();
         this.addRenderableWidget(saveButton);
         this.addRenderableWidget(cancelButton);
     }
 
-    private void save() {
-        int interval = 0;
-        try {
-            interval = Integer.parseInt(this.intervalWidget.getValue());
-            this.intervalError = false;
-        } catch (Exception e) {
-            this.intervalError = true;
-        }
+    private void save(Button button) {
+        int interval = getValueFromEditBox(this.intervalWidget, (bool) -> this.intervalError = bool);
+
         if (this.intervalError) {
             return;
         }
@@ -65,26 +56,39 @@ public class SequencerScreen extends Screen {
         this.onClose();
     }
 
-    private void cancel() {
-        this.onClose();
-    }
-
     @Override
     public void render(GuiGraphics graphics, int mouseX, int mouseY, float partialTicks) {
         this.renderBackground(graphics);
-        this.intervalWidget.render(graphics, mouseX, mouseY, partialTicks);
         super.render(graphics, mouseX, mouseY, partialTicks);
 
-        graphics.drawCenteredString(this.font, Component.translatable("screen.additionalredstone.sequencer.description"),
-                this.width / 2, 30, Constants.TEXT_COLOR_SCREEN);
+        // render title
+        graphics.drawString(this.font, this.title, this.leftPos + 9, this.topPos + 6, 4210752, false);
 
-        graphics.drawString(this.font, Component.translatable("screen.additionalredstone.sequencer.interval"),
-                this.width / 2 - 130, 65, Constants.TEXT_COLOR_SCREEN);
+        // render description for the edit boxes
+        graphics.drawString(this.font, INTERVAL, this.leftPos + 9, this.topPos + 30, 4210752, false);
 
+        // render red cross next to the edit box
         if (this.intervalError) {
-            graphics.blit(ERROR, this.width / 2 + 140, 61, 0, 0, 18, 18, 18, 18);
+            this.renderErrorIcon(graphics, this.leftPos + 227, this.topPos + 25);
         }
 
+        // render tooltips over edit boxes when focused
+        if (this.intervalWidget.isMouseOver(mouseX, mouseY)) {
+            graphics.renderTooltip(this.font, CustomScreen.TICK_DESCRIPTION, mouseX, mouseY);
+        }
     }
 
+    @Override
+    public void renderBackground(GuiGraphics graphics) {
+        super.renderBackground(graphics);
+        this.renderTexture(graphics, TEXTURE);
+    }
+
+    @Override
+    public boolean keyPressed(int $$0, int $$1, int $$2) {
+        if (this.intervalWidget.isFocused()) {
+            this.intervalError = false;
+        }
+        return super.keyPressed($$0, $$1, $$2);
+    }
 }