+++ Common/src/main/java/com/tristankechlo/additionalredstone/client/screen/TimerScreen.java
@@ -2,85 +2,79 @@ package com.tristankechlo.additionalredstone.client.screen;
 
 import com.tristankechlo.additionalredstone.Constants;
 import com.tristankechlo.additionalredstone.blockentity.TimerBlockEntity;
+import com.tristankechlo.additionalredstone.client.util.CustomScreen;
+import com.tristankechlo.additionalredstone.init.ModBlocks;
 import com.tristankechlo.additionalredstone.network.IPacketHandler;
+import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.components.EditBox;
-import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
 
 import java.time.LocalTime;
+import java.util.function.Consumer;
 
-public class TimerScreen extends Screen {
+public class TimerScreen extends CustomScreen {
 
-    private static final Component TITLE = Component.translatable("screen.additionalredstone.oscillator");
-    private static final ResourceLocation ERROR = new ResourceLocation(Constants.MOD_ID, "textures/gui/icons.png");
+    private static final Component TITLE = ModBlocks.TIMER_BLOCK.get().getName().withStyle(ChatFormatting.BOLD);
+    private static final ResourceLocation TEXTURE = new ResourceLocation(Constants.MOD_ID, "textures/gui/timer_screen.png");
+    private static final Component POWER_ON = Component.translatable("screen.additionalredstone.timer.power.on");
+    private static final Component POWER_OFF = Component.translatable("screen.additionalredstone.timer.power.off");
+    private static final Component INTERVAL = Component.translatable("screen.additionalredstone.timer.interval");
+    private static final Component DESCRIPTION = Component.translatable("screen.additionalredstone.timer.description");
     private EditBox powerUpWidget;
     private EditBox powerDownWidget;
     private EditBox intervalWidget;
-    private Button saveButton;
-    private Button cancelButton;
-    private BlockPos pos;
-    private int powerUpTime;
-    private int powerDownTime;
-    private int interval;
+    private final BlockPos pos;
+    private final int initialPowerUpTime;
+    private final int initialPowerDownTime;
+    private final int initialInterval;
     private boolean powerUpError = false;
     private boolean powerDownError = false;
     private boolean intervalError = false;
 
     public TimerScreen(int powerUp, int powerDown, int interval, BlockPos pos) {
-        super(TITLE);
-        this.powerUpTime = powerUp;
-        this.powerDownTime = powerDown;
-        this.interval = interval;
+        super(TITLE, 256, 152);
+        this.initialPowerUpTime = powerUp;
+        this.initialPowerDownTime = powerDown;
+        this.initialInterval = interval;
         this.pos = pos;
-    }
-
-    @Override
-    public void tick() {
-        this.powerUpWidget.tick();
-        this.powerDownWidget.tick();
-        this.intervalWidget.tick();
-    }
-
-    @Override
-    public boolean isPauseScreen() {
-        return false;
     }
 
     @Override
     protected void init() {
         super.init();
-        this.powerUpWidget = new EditBox(this.font, this.width / 2 + 32, 60, 98, 20, Component.translatable("screen.additionalredstone.timer.power.on"));
-        this.powerDownWidget = new EditBox(this.font, this.width / 2 + 32, 90, 98, 20, Component.translatable("screen.additionalredstone.timer.power.off"));
-        this.intervalWidget = new EditBox(this.font, this.width / 2 + 32, 120, 98, 20, Component.translatable("screen.additionalredstone.timer.interval"));
-        this.addWidget(this.powerUpWidget);
-        this.addWidget(this.powerDownWidget);
-        this.addWidget(this.intervalWidget);
-        this.powerUpWidget.setMaxLength(5);
-        this.powerDownWidget.setMaxLength(5);
-        this.intervalWidget.setMaxLength(5);
-        this.setInitialFocus(this.powerUpWidget);
-        this.powerUpWidget.setFocused(true);
-        this.powerUpWidget.setValue(String.valueOf(this.powerUpTime));
-        this.powerDownWidget.setValue(String.valueOf(this.powerDownTime));
-        this.intervalWidget.setValue(String.valueOf(this.interval));
+        this.powerUpWidget = new EditBox(this.font, this.leftPos + 176, this.topPos + 24, 70, 20, POWER_ON);
+        this.powerDownWidget = new EditBox(this.font, this.leftPos + 176, this.topPos + 57, 70, 20, POWER_OFF);
+        this.intervalWidget = new EditBox(this.font, this.leftPos + 176, this.topPos + 90, 70, 20, INTERVAL);
+        this.powerUpWidget.setMaxLength(10);
+        this.powerDownWidget.setMaxLength(10);
+        this.intervalWidget.setMaxLength(10);
+        this.powerUpWidget.setValue(String.valueOf(this.initialPowerUpTime));
+        this.powerDownWidget.setValue(String.valueOf(this.initialPowerDownTime));
+        this.intervalWidget.setValue(String.valueOf(this.initialInterval));
+        this.addRenderableWidget(this.powerUpWidget);
+        this.addRenderableWidget(this.powerDownWidget);
+        this.addRenderableWidget(this.intervalWidget);
 
-        this.saveButton = new Button.Builder(Component.translatable("screen.additionalredstone.save"), (b) -> this.save())
-                .pos(this.width / 2 - 110, 160).size(100, 20).build();
-        this.cancelButton = new Button.Builder(Component.translatable("screen.additionalredstone.cancel"), (b) -> this.cancel())
-                .pos(this.width / 2 + 10, 160).size(100, 20).build();
+        Button saveButton = new Button.Builder(CustomScreen.TEXT_SAVE, this::save)
+                .pos(this.leftPos + 9, this.topPos + 123).size(116, 20)
+                .tooltip(CustomScreen.TOOLTIP_SAVE.get()).build();
+        Button cancelButton = new Button.Builder(CustomScreen.TEXT_CANCEL, (b) -> this.onClose())
+                .pos(this.leftPos + 131, this.topPos + 123).size(116, 20)
+                .tooltip(CustomScreen.TOOLTIP_CANCEL.get()).build();
         this.addRenderableWidget(saveButton);
         this.addRenderableWidget(cancelButton);
     }
 
-    private void save() {
-        int powerUp = this.getPowerUpTime();
-        int powerDown = this.getPowerDownTime();
-        int interval = this.getInterval();
+    private void save(Button button) {
+        int powerUp = getTimeFromEditBox(this.powerUpWidget, (b) -> this.powerUpError = b);
+        int powerDown = getTimeFromEditBox(this.powerDownWidget, (b) -> this.powerDownError = b);
+        int interval = getValueFromEditBox(this.intervalWidget, (b) -> this.intervalError = b);
+
         if (this.powerUpError || this.powerDownError || this.intervalError) {
             return;
         }
@@ -88,94 +82,74 @@ public class TimerScreen extends Screen {
         this.onClose();
     }
 
-    private void cancel() {
-        this.onClose();
-    }
-
     @Override
     public void render(GuiGraphics graphics, int mouseX, int mouseY, float partialTicks) {
         this.renderBackground(graphics);
-        this.powerUpWidget.render(graphics, mouseX, mouseY, partialTicks);
-        this.powerDownWidget.render(graphics, mouseX, mouseY, partialTicks);
-        this.intervalWidget.render(graphics, mouseX, mouseY, partialTicks);
         super.render(graphics, mouseX, mouseY, partialTicks);
 
-        graphics.drawCenteredString(this.font, Component.translatable("screen.additionalredstone.timer.description"),
-                this.width / 2, 30, Constants.TEXT_COLOR_SCREEN);
+        // render title
+        graphics.drawString(this.font, this.title, this.leftPos + 9, this.topPos + 6, 4210752, false);
 
-        graphics.drawString(this.font, Component.translatable("screen.additionalredstone.timer.power.on"),
-                this.width / 2 - 130, 65, Constants.TEXT_COLOR_SCREEN);
-        graphics.drawString(this.font, Component.translatable("screen.additionalredstone.timer.power.off"),
-                this.width / 2 - 130, 95, Constants.TEXT_COLOR_SCREEN);
-        graphics.drawString(this.font, Component.translatable("screen.additionalredstone.timer.interval"),
-                this.width / 2 - 130, 125, Constants.TEXT_COLOR_SCREEN);
+        // render description for the edit boxes
+        graphics.drawString(this.font, POWER_ON, this.leftPos + 9, this.topPos + 30, 4210752, false);
+        graphics.drawString(this.font, POWER_OFF, this.leftPos + 9, this.topPos + 63, 4210752, false);
+        graphics.drawString(this.font, INTERVAL, this.leftPos + 9, this.topPos + 96, 4210752, false);
 
+        // render red cross next to the edit box
         if (this.powerUpError) {
-            graphics.blit(ERROR, this.width / 2 + 140, 61, 0, 0, 18, 18, 18, 18);
+            this.renderErrorIcon(graphics, this.leftPos + 227, this.topPos + 25);
         }
         if (this.powerDownError) {
-            graphics.blit(ERROR, this.width / 2 + 140, 91, 0, 0, 18, 18, 18, 18);
+            this.renderErrorIcon(graphics, this.leftPos + 227, this.topPos + 58);
         }
         if (this.intervalError) {
-            graphics.blit(ERROR, this.width / 2 + 140, 121, 0, 0, 18, 18, 18, 18);
+            this.renderErrorIcon(graphics, this.leftPos + 227, this.topPos + 91);
+        }
+
+        // render tooltips over edit boxes when focused
+        if (this.intervalWidget.isMouseOver(mouseX, mouseY)) {
+            graphics.renderTooltip(this.font, CustomScreen.TICK_DESCRIPTION, mouseX, mouseY);
+        }
+        if (this.powerUpWidget.isMouseOver(mouseX, mouseY) || this.powerDownWidget.isMouseOver(mouseX, mouseY)) {
+            graphics.renderTooltip(this.font, DESCRIPTION, mouseX, mouseY);
         }
     }
 
-    private int getPowerUpTime() {
-        int powerUp = 0;
-        String text = this.powerUpWidget.getValue();
-        if (text.equalsIgnoreCase("24:00")) {
-            text = "00:00";
-        }
-        try {
-            LocalTime time = LocalTime.parse(text);
-            powerUp = (time.getHour() * 1000) + (int) (time.getMinute() * 16.67) - 6000;
-            if (powerUp < 0) {
-                powerUp = 24000 + powerUp;
-            }
+    @Override
+    public void renderBackground(GuiGraphics graphics) {
+        super.renderBackground(graphics);
+        this.renderTexture(graphics, TEXTURE);
+    }
+
+    @Override
+    public boolean keyPressed(int $$0, int $$1, int $$2) {
+        if (this.powerUpWidget.isFocused()) {
             this.powerUpError = false;
-        } catch (Exception e) {
-            try {
-                powerUp = Integer.parseInt(text);
-            } catch (Exception ex) {
-                this.powerUpError = true;
-            }
         }
-        return Mth.clamp(powerUp, TimerBlockEntity.minTime, TimerBlockEntity.maxTime);
+        return super.keyPressed($$0, $$1, $$2);
     }
 
-    private int getPowerDownTime() {
-        int powerDown = 0;
-        String text = this.powerDownWidget.getValue();
+    private static int getTimeFromEditBox(EditBox widget, Consumer<Boolean> consumer) {
+        int returnTime = 0;
+        String text = widget.getValue();
         if (text.equalsIgnoreCase("24:00")) {
             text = "00:00";
         }
         try {
             LocalTime time = LocalTime.parse(text);
-            powerDown = (time.getHour() * 1000) + (int) (time.getMinute() * 16.67) - 6000;
-            if (powerDown < 0) {
-                powerDown = 24000 + powerDown;
+            returnTime = (time.getHour() * 1000) + (int) (time.getMinute() * 16.67) - 6000;
+            if (returnTime < 0) {
+                returnTime = 24000 + returnTime;
             }
-            this.powerDownError = false;
+            consumer.accept(false);
         } catch (Exception e) {
             try {
-                powerDown = Integer.parseInt(text);
+                returnTime = Integer.parseInt(text);
             } catch (Exception ex) {
-                this.powerDownError = true;
+                consumer.accept(true);
             }
         }
-        return Mth.clamp(powerDown, TimerBlockEntity.minTime, TimerBlockEntity.maxTime);
-    }
-
-    private int getInterval() {
-        int interval = 0;
-        try {
-            interval = Integer.parseInt(this.intervalWidget.getValue());
-            this.intervalError = false;
-        } catch (Exception e) {
-            this.intervalError = true;
-        }
-        return Mth.clamp(interval, 1, 1000);
+        return Mth.clamp(returnTime, TimerBlockEntity.minTime, TimerBlockEntity.maxTime);
     }
 
 }